Add WP-CLI commands for doing API requests
Adds following WP-CLI commands:
- wp seravo api get site          # This is /site/xxx
- wp seravo api get shadow        # This is /site/xxx/shadows
- wp seravo api get shadow [name] # This is /site/xxx/shadow/xxx

Commands can be used for testing the API responses. The responses
are in Seravo Plugin format meaning that the responses have been
modified to be Seravo Plugin compatible.

All the commands accept --version=xx for specifying the API
version to use. The specified version might not be used if it's invalid.